remove Game from src/index.js
moving the Game component into features\tic-tac-toe
