+++ front-end/src/features/tic-tac-toe/index.js
@@ -1,0 +1,82 @@
+import React from 'react';
+
+function Square(props) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
+
+class Board extends React.PureComponent {
+  renderSquare(i) {
+    return <Square value={this.props.squares[i]} onClick={() => this.props.onClick(i)} />;
+  }
+
+  render() {
+    return (
+      <div>
+        <div className="status">{status}</div>
+        <div className="board-row">
+          {this.renderSquare(0)}
+          {this.renderSquare(1)}
+          {this.renderSquare(2)}
+        </div>
+        <div className="board-row">
+          {this.renderSquare(3)}
+          {this.renderSquare(4)}
+          {this.renderSquare(5)}
+        </div>
+        <div className="board-row">
+          {this.renderSquare(6)}
+          {this.renderSquare(7)}
+          {this.renderSquare(8)}
+        </div>
+      </div>
+    );
+  }
+}
+
+export class Game extends React.PureComponent {
+  jumpTo(step) {
+    this.props.dispatch({ type: 'JUMP_TO_STEP', step });
+  }
+
+  reset() {
+    this.props.dispatch({ type: 'RESET' });
+  }
+
+  handleClick(boxIndex) {
+    this.props.dispatch({ type: 'PERFORM_MOVE', boxIndex: boxIndex });
+  }
+
+  render() {
+    const { history, stepNumber, xIsNext, winner } = this.props.gameState;
+    const current = history[stepNumber];
+    const status = winner ? 'Winner: ' + winner : 'Next player: ' + (xIsNext ? 'X' : 'O');
+
+    const moves = history.map((step, move) => {
+      const desc = move ? 'Go to move #' + move : 'Go to game start';
+      return (
+        <li key={move}>
+          <button onClick={() => this.jumpTo(move)}>{desc}</button>
+        </li>
+      );
+    });
+
+    return (
+      <div className="game">
+        <div className="game-board">
+          <Board squares={current.squares} onClick={(i) => this.handleClick(i)} />
+        </div>
+        <div className="game-info">
+          <div>
+            <button onClick={() => this.reset()}>Reset the Game</button>
+          </div>
+          <div>{status}</div>
+          <ol>{moves}</ol>
+        </div>
+      </div>
+    );
+  }
+}
